+++ allanoguis/package.json
@@ -15,6 +15,7 @@
     "framer-motion": "^11.5.4",
     "lodash": "^4.17.21",
     "next": "^14.2.11",
+    "next-themes": "^0.3.0",
     "nextjs": "^0.0.3",
     "react": "^18.3.1",
     "react-dom": "^18",
