+++ allanoguis/src/components/navbar/HeaderMain.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import Link from "next/link";
 import { useSelectedLayoutSegment } from "next/navigation";
-import useScroll from "@/hooks/use-scroll";
+import useScroll from "@/hooks/useScroll";
 import { cn } from "@/lib/utils";
 import { WebhookIcon } from "@primer/octicons-react";
 import FullscreenMode from "@/lib/useFullscreen";
